requirements.txt
Django==2.2
django-allauth==0.41.0
django-crispy-forms==1.9.0
Pillow==7.1.1
stripe==2.46.0
sendgrid==6.3.0
python-decouple==3.3